fix: API

+++ Hotel.WebApi/Dtos/Habitacion/HabitacionDtoBase.cs
@@ -4,7 +4,7 @@ public class HabitacionDtoBase : DtoBase
 {
     public string? Numero { get; set; }
     public string? Detalle { get; set; }
-    public float? Precio { get; set; }
+    public decimal? Precio { get; set; }
     public int IdEstadoHabitacion { get; set; }
     public int IdPiso { get; set; }
     public int IdCategoria { get; set; }
